refactor some services. fix tests
