+++ src/app/alarm/alarm-page.component.spec.ts
@@ -2,6 +2,10 @@ import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 import { IonicModule } from '@ionic/angular';
 
 import { AlarmPage } from './alarm-page.component';
+import {HomeBrokerService} from '../services/home-broker.service';
+import {HomeBrokerServiceMock} from '../services/home-broker.service.mock';
+import {CommandRequestService} from '../services/command-request.service';
+import {CommandRequestServiceMock} from '../services/command-request.service.mock';
 
 describe('AlarmPage', () => {
   let component: AlarmPage;
@@ -9,6 +13,10 @@ describe('AlarmPage', () => {
 
   beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
+      providers: [
+        { provide: HomeBrokerService, useClass: HomeBrokerServiceMock },
+        { provide: CommandRequestService, useClass: CommandRequestServiceMock }
+      ],
       declarations: [AlarmPage],
       imports: [IonicModule.forRoot()]
     }).compileComponents();
