+++ src/app/gate/gate-page.component.spec.ts
@@ -2,6 +2,7 @@ import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 import { IonicModule } from '@ionic/angular';
 
 import { GatePage } from './gate-page.component';
+import {RouterTestingModule} from "@angular/router/testing";
 
 describe('GatePage', () => {
   let component: GatePage;
@@ -10,7 +11,7 @@ describe('GatePage', () => {
   beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [GatePage],
-      imports: [IonicModule.forRoot()]
+      imports: [RouterTestingModule, IonicModule.forRoot()]
     }).compileComponents();
 
     fixture = TestBed.createComponent(GatePage);
