+++ src/app/info/info-page.component.spec.ts
@@ -2,6 +2,9 @@ import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 import { IonicModule } from '@ionic/angular';
 
 import { InfoPage } from './info-page.component';
+import {HomeBrokerService} from "../services/home-broker.service";
+import {HomeBrokerServiceMock} from "../services/home-broker.service.mock";
+import {RouterTestingModule} from "@angular/router/testing";
 
 describe('InfoPage', () => {
   let component: InfoPage;
@@ -9,8 +12,14 @@ describe('InfoPage', () => {
 
   beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
+      providers: [
+        {provide: HomeBrokerService, useClass: HomeBrokerServiceMock}
+      ],
       declarations: [InfoPage],
-      imports: [IonicModule.forRoot()]
+      imports: [
+        RouterTestingModule,
+        IonicModule.forRoot()
+      ]
     }).compileComponents();
 
     fixture = TestBed.createComponent(InfoPage);
