+++ src/app/services/command-request.service.mock.ts
@@ -1,0 +1,9 @@
+import {CommandRequestService} from './command-request.service';
+
+export class CommandRequestServiceMock extends CommandRequestService {
+  sendAlarmEcuSet(state: number): void {
+  }
+
+  sendAlarmEcuToggle(): void {
+  }
+}
